phase2/requirements.txt
# Phase 2 - Dépendances avec Kafka + RAG
# Installation: pip install -r requirements.txt

# LLM Framework
anthropic>=0.40.0

# Data Validation
pydantic>=2.5.0

# Environment
python-dotenv>=1.0.0

# Kafka
confluent-kafka>=2.3.0

# RAG & Vector Store
chromadb>=0.4.22
sentence-transformers>=2.2.0
